+++ Wire-iOS Tests/ConversationList/Container/ConversationListViewControllerViewModelSnapshotTests.swift
@@ -66,69 +66,3 @@ final class ConversationListViewControllerViewModelSnapshotTests: CoreDataSnapsh
         }
     }
 }
-
-final class ConversationActionControllerSnapshotTests: ZMSnapshotTestCase {
-    var viewModel: ConversationListViewController.ViewModel!
-    fileprivate var mockViewController: MockConversationListContainer!
-    
-    override func setUp() {
-        super.setUp()
-        
-        let account = Account.mockAccount(imageData: Data())
-        viewModel = ConversationListViewController.ViewModel(account: account, selfUser: MockUser.mockSelf())
-        
-        mockViewController = MockConversationListContainer(viewModel: viewModel)
-        
-        viewModel.viewController = mockViewController
-        
-        recordMode = true
-    }
-    
-    override func tearDown() {
-        viewModel = nil
-        mockViewController = nil
-        
-        super.tearDown()
-    }
-
-
-    func testForActionMenu_removeFolder() {
-        let mockConversation = MockConversation()
-        mockConversation.folderName = "Test Folder"
-        viewModel.showActionMenu(for: mockConversation, from: mockViewController.view)
-
-        let sut = (viewModel?.actionsController)!
-
-        verifyAlertController(sut.alertController!)
-    }
-}
-
-fileprivate final class MockConversation: ConversationInterface {
-    var conversationType: ZMConversationType = .group
-
-    var teamRemoteIdentifier: UUID?
-
-    var connectedUser: ZMUser?
-
-    var displayName: String = "Mock Conversation"
-
-    var isArchived: Bool = false
-
-    var isReadOnly: Bool = false
-
-    var isFavorite: Bool = false
-
-    var mutedMessageTypes: MutedMessageTypes = .none
-
-    var activeParticipants: Set<ZMUser> = []
-
-    var folderName: String?
-
-    var unreadMessages: [ZMConversationMessage] = []
-
-    func canMarkAsUnread() -> Bool {
-        return true
-    }
-
-
-}
